+++ context/app/static/js/components/detailPage/StatusIcon/ColoredStatusIcon/style.ts
@@ -1,20 +1,7 @@
 import React from 'react';
 import { styled } from '@mui/material/styles';
-import { InfoRounded, ErrorRounded, WarningRounded, CheckCircleRounded } from '@mui/icons-material';
 import { SvgIconProps } from '@mui/material/SvgIcon';
-
-export interface ColoredStatusIconProps extends SvgIconProps {
-  $iconStatus: 'info' | 'success' | 'warning' | 'error';
-}
-
-export const iconSymbolStatusMap: {
-  [key in ColoredStatusIconProps['$iconStatus']]: React.ComponentType<SvgIconProps>;
-} = {
-  info: InfoRounded,
-  success: CheckCircleRounded,
-  warning: WarningRounded,
-  error: ErrorRounded,
-};
+import { ColoredStatusIconProps } from 'js/shared-styles/icons/SeverityIcon';
 
 export const getStyledIcon = (
   IconComponent: React.ComponentType<SvgIconProps>,
